fix: update module path and fix publisher example
- Update module path from github.com/nano-interactive/go-amqp to github.com/nano-interactive/go-amqp/v3
- Fix publisher example to use correct connection configuration and exchange declaration
- Remove self-dependency from go.mod
- Update import paths to use v3 version

+++ connection/connection.go
@@ -253,7 +253,7 @@ func (c *Connection) connect() func(ctx context.Context) error {
 		config := amqp091.Config{
 			SASL:       nil,
 			Vhost:      c.config.Vhost,
-			ChannelMax: c.config.Channels,
+			ChannelMax: uint16(c.config.Channels),
 			FrameSize:  c.config.FrameSize,
 			Heartbeat:  3 * time.Second,
 			Properties: properties,
